Vista para editar Producto finalizado.

+++ database/migrations/2019_08_28_113847_create_materia_primas_table.php
@@ -17,13 +17,13 @@ class CreateMateriaPrimasTable extends Migration
             $table->bigIncrements('id');
             $table->string('nombre')->nullable()->unique();
             $table->string('detalle')->nullable();
-            $table->integer('cantidad')->nullable();
+            $table->bigInteger('cantidad')->nullable();
             $table->double('precioUnitario')->nullable();
             $table->string('color')->nullable();
             $table->string('imagenPrincipal')->nullable();
             $table->softDeletes();
-           
-            
+
+
             $table->timestamps();
         });
     }
@@ -39,7 +39,6 @@ class CreateMateriaPrimasTable extends Migration
             $table->dropColumn('id');
             $table->dropColumn('nombre');
             $table->dropColumn('detalle');
-            
         });
     }
 }
